fix: avoid multiple refreshToken calls
